Update exercise description
In the seminar yesterday I could feel that people are confused about the
exercise and that it's also too much for just one exercise. So I decided to make
middlewares optional and authorization will become part of an upcoming exercise.
I also decided to extend the deadline by one week.

Those groups who did all of the previous tasks already will have less work
to do in the following exercises.