Write first draft of the exam
My plan is to publish a couple of exam questions so our students can
have a look and ask questions if some exercises descriptions are hard to
understand. I would like to minimize the number of unpleasant surprises
in the final exam.

In the final version I will vary the content of some questions, remove questions
and add a couple of new questions. The general structure should stay roughly the
same.

I've used
https://de.overleaf.com/latex/templates/taif-university-exam-template/rpqzthxhtnsh
and
https://de.wikipedia.org/wiki/Datei:HTW_Berlin_logo.svg